+++ gitstat/__init__.py
@@ -1,2 +1,2 @@
-VERSION_INFO = (0, 3, 0)
+VERSION_INFO = (0, 4, 0)
 VERSION = '.'.join(str(c) for c in VERSION_INFO)